apply suggestions from @43081j's review
